requirements.txt
#basics
pytorch-lightning==1.7.7
torchmetrics==0.10.0
einops==0.5.0
editdistance==0.6.0
pillow==9.2.0
# for pytorch-lightning CLI
jsonargparse[signatures]==4.15.2
# dev-dependency
flake8==5.0.4
black==22.10.0
isort==5.10.1
jupyter==1.0.0
opencv-python==4.6.0.66
matplotlib==3.6.1
# for test in crohme
typer==0.6.1
beautifulsoup4==4.11.1
lxml>=4.9.1
numpy>=1.23.1
levenshtein>=0.20.8